… radio/src/tests/fonts_480x272.png
… radio/src/tests/transparency_480x272.png