This are the baseline tests for #4508. After the optimization of graphics code they can be used again to determine the improvement.

Current results on beta Horus, compiled with:
```-DCLI=YES -DDEBUG=YES -DCMAKE_BUILD_TYPE=Debug```

```
> test memspd
Starting memory speed test...
Test testMemoryReadFrom_RAM_8bit speed: 6730.77, (1400 runs in 208 ms)
Test testMemoryReadFrom_RAM_32bit speed: 27000.00, (5400 runs in 200 ms)
Test testMemoryReadFrom_SDRAM_8bit speed: 1428.57, (300 runs in 210 ms)
Test testMemoryReadFrom_SDRAM_32bit speed: 5092.59, (1100 runs in 216 ms)
Test testMemoryCopyFrom_RAM_to_SDRAM_8bit speed: 4500.00, (900 runs in 200 ms)
Test testMemoryCopyFrom_RAM_to_SDRAM_32bit speed: 14705.88, (3000 runs in 204 ms)
Test testMemoryCopyFrom_SDRAM_to_SDRAM_8bit speed: 5092.59, (1100 runs in 216 ms)
Test testMemoryCopyFrom_SDRAM_to_SDRAM_32bit speed: 13500.00, (2700 runs in 200 ms)
Disabling LCD...
Test testMemoryReadFrom_RAM_8bit speed: 6796.12, (1400 runs in 206 ms)
Test testMemoryReadFrom_RAM_32bit speed: 27000.00, (5400 runs in 200 ms)
Test testMemoryReadFrom_SDRAM_8bit speed: 2521.01, (600 runs in 238 ms)
Test testMemoryReadFrom_SDRAM_32bit speed: 8910.89, (1800 runs in 202 ms)
Test testMemoryCopyFrom_RAM_to_SDRAM_8bit speed: 4500.00, (900 runs in 200 ms)
Test testMemoryCopyFrom_RAM_to_SDRAM_32bit speed: 14705.88, (3000 runs in 204 ms)
Test testMemoryCopyFrom_SDRAM_to_SDRAM_8bit speed: 5092.59, (1100 runs in 216 ms)
Test testMemoryCopyFrom_SDRAM_to_SDRAM_32bit speed: 13500.00, (2700 runs in 200 ms)
Total speed: 161076.92
```

```
>test graphics
Starting graphics performance test...
Test testEmpty speed: 17823202.00, (17823200 runs in 1000 ms)
Test testDrawSolidHorizontalLine2 speed: 79000.00, (79000 runs in 1000 ms)
Test testDrawSolidVerticalLine2 speed: 34331.34, (34400 runs in 1002 ms)
Test testDrawDiagonalLine speed: 4282.87, (4300 runs in 1004 ms)
Test testDrawSolidFilledRectangle speed: 15.76, (100 runs in 6346 ms)
Test testDrawSolidFilledRoundedRectangle speed: 62.89, (100 runs in 1590 ms)
Test testDrawRect speed: 662.88, (700 runs in 1056 ms)
Test testDrawFilledRectangle speed: 30.03, (100 runs in 3330 ms)
Test testDrawBlackOverlay speed: 9.65, (100 runs in 10368 ms)
Test testDrawText speed: 469.48, (600 runs in 1278 ms)
Test testDrawTextVertical speed: 627.24, (700 runs in 1116 ms)
Test testClear speed: 338.98, (400 runs in 1180 ms)
Total speed: 119831.16
```